requirements.txt
cffi==1.15.1
cryptography==40.0.2
dnspython==2.3.0
pyjwt==2.6.0
pycparser==2.21
pymongo==4.3.3
PyYAML==6.0
tornado==6.3
